move user router to users.js file + clean up error
