+++ routes/users.js
@@ -1,17 +1,66 @@
 var express = require('express');
-var router = express.Router();
+var userRouter = express.Router();
 
-/* GET users listing. */
-router.get('/', function(req, res) {
-  res.send('respond with a resource');
+userRouter.use(isAuthenticated)
+
+userRouter.get("/", function(req, res, next) {
+  req.user.isAdmin(function(admin) {
+    res.render("room", {
+      username: req.user.username,
+      admin: admin
+    });
+  });
 });
 
-router.get('/sudo', function(req, res) {
-  // TODO: redo user object
-  if (user.isAdmin()) {
+userRouter.get("/changepassword", function(req, res, next) {
+  res.render("passChange", {
+    username: req.user.username
+  });
+});
 
+userRouter.post("/changepassword", function(req, res, next) {
+  req.user.changePassword(req.body.oldPassword, req.body.newPassword, function(err, data) {
+    if (err) return next(err);
+    if (data) {
+      res.send({
+        success: true
+      });
+    } else {
+      res.send({
+        success: false
+      });
+    }
+  });
+});
+
+userRouter.get("/admin", isAdmin, function(req, res, next) {
+  res.render("admin", {
+    username: req.user.username
+  });
+});
+
+
+module.exports = userRouter;
+
+
+function isAdmin(req, res, next) {
+  req.user.isAdmin(function(admin) {
+    if (admin) {
+      return next();
+    } else {
+      res.redirect("/user");
+    }
+  });
+}
+
+
+// TODO: this function should be moved or passed from one source
+// XXX: code duplicate in ./index.js
+function isAuthenticated(req, res, next) {
+  console.log("called")
+  if (req.isAuthenticated()) {
+    return next();
+  } else {
+    res.redirect("/"); // send the user to the landing page if he is not logged in...
   }
-  // TODO: sudo terminalmonly for admin users
-});
-
-module.exports = router;
+}
